Move action to connected component
